Cleanup type checker warnings/errors
I enabled PyLance's "basic" type checking, and found several
opportunities to make the code more obvious.
The main issues remaining is that logicvec.__getitem__ can return either
logic or logicvec, depending on whether the input argument is an index
(int or logicvec) or a slice.
This has the additional effect of making logicvec.__iter__ yield either
logic or logicvec.