Implement basic logicvec addition
The new logicvec add method implements ripple carry. This is a brute
force approach, but it 1) gives the right answer, and 2) allows
predictable X propagation.

The add method implements full addition, with carry in, carry out, and
overflow detection. The __add__, __sub__, and __neg__ methods, on the
other hand, ignore carry/overflow details, and just do fixed width
addition. This seems like the simplest possible choice.

If the user really wants to use the '+' operator while maintaining carry
out, he can widen the inputs first. Verilog actually does this in a
sneaky way - if the LHS of a statement has enough bits for the carry,
it expands the widths of the operands on the RHS to match. I don't think
that can work in Python, so no point figuring out whether it makes
sense.