Move lifted bool functions to PCN module
Rename the Boolean functions from {not, or, and, ...} to
{lnot, lnor, land, ...}.
This avoids collision with Python keywords, and introduces the
concept of "lifted" functions.

In "An Extensible SAT-solver" by Niklas Een and Niklas Sorensson,
they introduce the concept of a "lifted" Boolean.
This is a Boolean type that includes an encoding for "undefined".
It *lifts* Boolean values from 2 states to 3 states.
We are doing a similar thing here.
Our PCN logic function include semantics for propagating undefined
values for both NULL and "don't care" positional cube encodings.
Our Boolean functions are "lifted" to four states.

In addition to simply renaming functions, this change separates the
concerns of implementing the Python UI layer in logic.py,
and the basic math layer in pcn.py.
This makes it easier for us to move performance-critical code to C in
the future.