Remove Sim._prune method
Previously, we were removing both the edges and nodes from the graph
after the task is called.
Removing the edges is required, but removing the nodes is not.
Once a simulation is started, the set of variables, events, and
tasks is static. Repeatedly adding and removing them costs some cpu
time, and doesn't achieve anything.