Remove dependency on networkx
We were only using a fraction of its capabilities.
Replace the graph with two tables:
* var => {task, ...}
* (var, task) => trigger

The first is an adjacency list, to get a list of tasks that might need
to execute if a variable is touched. The second is a mapping from (var,
task) to a function that determines if the task is called.