Overhaul how lbool data types are used
Change the naming convention of lbool.vec: {X, 0, 1, W}.
We should prefer the strong version of X for uninitialized values.
There might yet be a need for a "don't care" value in some future truth
table implementation.

Change Bits so it is using lbool.vec under the hood.
First, we need to review the usage of multiple dimensions.
There's at least one known bug (in unit tests), the cat method's
'flatten' argument is confusing, and some other problems.
Second, we should probably make Bits a subclass of Vec.
It might ultimately be simpler that way.