Improve recursive select function
Improve one case that was stacking scalars into a vector. Just
create the vector more directly, which will speed things up.

This has the nice benefit that it gets rid of the weird usage of
shape=() for scalars. Only support (n,) now, where (0,) is empty,
(1,) is scalar, and (n,) is Vec[n].