Allow more than one event loop
Fix #15

Remove the sim.get_loop function, and replace it with
get/set/new_event_loop methods.

Create a new sim.now function, which returns the current simulation
time. If there is no running loop, it raises a RuntimeError.

Update how simulations are started. Create new sim.run and sim.irun (for
iteration) functions. These functions will create a new event loop is
none is provided. It's still possible to provide a time limit, and
continue the simulation.

The EventLoop class no longer needs a list of "initial" tasks. Those are
provided as inputs to the run/irun functions.

Lots of little test updates for this API change.

This ended up being a large refactor, so probably forgot to mention a
couple things.